Change code as per review
