+++ common/api/serializers.py
@@ -29,3 +29,20 @@ class PublishedSerializer(serializers.ModelSerializer):
             "is_published",
             "publish_date",
         )
+
+
+class DynamicFieldsCategorySerializer():
+    def __init__(self, *args, **kwargs):
+        # Don't pass the 'fields' arg up to the superclass
+        fields = kwargs.pop('fields', None)
+
+        # Instantiate the superclass normally
+        super().__init__(*args, **kwargs)
+
+        if fields is not None:
+            # Drop any fields that are not specified in the `fields` argument.
+            allowed = set(fields)
+            existing = set(self.fields)
+            for field_name in existing - allowed:
+                self.fields.pop(field_name)
+
